ARCHIBALD - changed format of list of lines, fixed csv output, flipped lines
lines is now a list of pins with the associated colour, allows better flexibility and clarity for csv output. can be easily converted to xy with Canvas.Coords[pin]
or to Angles with np.linspace[0,2*pi,numPins],

list of lines are now flipped so that first lines found are drawn last.